+++ book_program.rb
@@ -1,33 +1,34 @@
 require 'open-uri'
 require 'digest'
 
-def baixar_livro(url)
-  morro_dos_ventos_uivantes = Digest::MD5.hexdigest(url)
+def download_book(url)
+  book_file = Digest::MD5.hexdigest(url)
 
-  if File.exist?(morro_dos_ventos_uivantes)
+  if File.exist?(book_file)
     puts "Recuperando arquivo do cache..."
-    livro = File.read(morro_dos_ventos_uivantes)
+    book = File.read(book_file)
   else
     puts "Baixando arquivo do livro..."
-    livro = URI.open(url).read
-    File.write(morro_dos_ventos_uivantes, livro)
+    book = URI.open(url).read
+    File.write(book_file, book)
   end
 
-  return livro
+  return book
 end
 
+def count_words(text, exclude_words = [])
+  words = text.downcase.gsub(/[^a-z\s]/, '').split(" ")
+  words = words.reject { |word| exclude_words.include?(word) }
+
+  frequency = words.group_by(&:itself).transform_values(&:count).sort_by { |word, count| -count }
+  frequency.first(20)
+end
 url = "https://www.gutenberg.org/cache/epub/768/pg768.txt"
-livro = baixar_livro(url)
+book = download_book(url)
 
-livro = livro.gsub(/[\n\r]/, " ")
-livro = livro.gsub(/[^a-zA-Z ]/, "")
-livro = livro.downcase
-contagem = Hash.new(0)
-livro.split.each { |palavra| contagem[palavra] += 1 }
+book_content = download_book(url)
+exclude_words = ["the", "and", "or", "a", "an", "but"]
+frequency = count_words(book_content, exclude_words)
 
-palavras_mais_utilizadas = contagem.sort_by { |_, count| -count }.first(20)
-
-puts "As 20 palavras mais utilizadas no livro (excluindo 'the', 'and', 'or', 'a', 'an' e 'but') são:"
-palavras_mais_utilizadas.each do |palavra, frequencia|
-  puts "#{palavra}: #{frequencia} vezes"
-end
+puts "As 20 palavras mais comuns no livro (excluindo #{exclude_words.join(', ')}) são:"
+frequency.each { |word, count| puts "#{word}: #{count}" }