+++ palíndromo.rb
@@ -1,17 +1,12 @@
 loop do
 
   print "Digite a primeira palavra (ou 'quit' para sair): "
-  word1 = gets.chomp.downcase
+  word1 = gets.chomp.strip.downcase
 
-  break if word1 == 'quit'
+ print "Digite a segunda palavra: "
+  word2 = gets.chomp.strip.downcase
 
-  print "Digite a segunda palavra: "
-  word2 = gets.chomp.downcase
-
-  if word2.empty?
-    puts "Erro: a segunda palavra não pode ser vazia."
-    next
-  end
+break if word1 == 'quit' && word2.empty?
 
   if word1 == word2.reverse
     puts "As palavras são palíndromos uma da outra."